apps/unautomated/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pytrends==4.9.2
pandas==2.1.3
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
aioredis==2.0.1
